+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added support for newer Ubuntu
 * Detect when a container for an old distribution cannot be started due to
   group v1 guest on v2 host (see https://github.com/lxc/lxc/issues/4072)
+* Use mmdebstrap to bootstrap Debian, if available
 
 # Version 0.21